Components: Extract the WritingFlow component as a reusable Editor component
